Fix regression in -x language handling in PNaCl driver and add test.
NaCl rev 11178 broke the -x handling, because it changed
the argument structure for the @SimpleCache wrapper.  It
changed dictionary keys from 'c' to ('c',), so the manual
setting of the cache entries from the ForceFileType function 
wasn't matching the lookup.  Fix that to match.

Fix "-x none" so that it actually works.

Add a test that the -x works, so that we don't regress.
Simplify the output capture by using driver_log instead.

BUG=http://code.google.com/p/chromium/issues/detail?id=242170

Review URL: https://chromiumcodereview.appspot.com/15627002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11385 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2