Add missing declaration for lstat() to <sys/stat.h>
Without this, building Python with nacl-newlib fails:

../Modules/posixmodule.c:6242:50: error: use of undeclared identifier 'lstat'
    return posix_do_stat(self, args, "et:lstat", lstat, NULL, NULL);

Python has an autoconf check for the presence of lstat() at link time.
When linking against libnosys, this autoconf check passes, because
libnosys contains a stub for lstat(), but taking the address of
lstat() fails because of the lack of a function declaration.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=1229
TEST=none

Review URL: https://codereview.chromium.org/18816004

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11693 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2