Tighten asm constraint in ARM platform qualification for unaligned
The ARM platform qualification test for unaligned access using LDRD
instructions triggered an assembler warning because it was emitting a
case that is specified as UNPREDICTABLE in the ARM processor manual.
To preclude this situation, the pair of destination registers in the
load must not overlap with either of the registers used in the
address calculation.  Adding the & modifier to the asm constraint on
the output operand expresses that it cannot involve any of the same
registers used in the input operands, which avoids the bad situation.

BUG= none
TEST= trybots
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/22464006

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@11939 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2