Bump archived pnacl frontend and bitcode tests to 11986.
Some of c11 and c++11 scons tests are currently failing.
These depend on newer headers:

errors about ATOMIC_BOOL_LOCK_FREE
errors about no member named 'thread' in namespace 'std'

As well as newer clang and LLVM intrinsics for:

__nacl_atomic_is_lock_free

For the header problem, I thought we could perhaps only
bring in an archived set of clang and LLVM binaries,
and let the headers change, but that doesn't really
reproduce the snapshot of bitcode.  It also doesn't
work because the newer headers require a newer
clang + LLVM to support them.

Perhaps in the future we would label each test as being
supported since NaCl revision N, vs M, and we'd maintain
a list of these revisions for archived frontends.  Or, we
would manually snapshot some bitcode, and everytime we add
a new test or modify a test to have more test cases, we
would need to create a snapshot for the new test by finding
an old archived frontend that supports that particular test,
but that sounds painful too.

BUG=none, red FYI bots
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/22914007

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12007 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2