[MIPS] Allow Trap instructions in untrusted code
MIPS needs to trap when divide by zero. This is done with the help of TRAP
instructions, so we allow them.
Minor changes had to be made to dgen_output.py as the code generated by the
previous version of the file does not satisfy code styles rules.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=2833
TEST= run_divide_by_zero_test

Review URL: https://chromiumcodereview.appspot.com/22464002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12009 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2