Allow debug stub to change memory-mapped data and code on Windows.
It is not possible to change access rights for memory mapped pages
as freely as for normal memory.

In order to read from memory-mapped region, we now only check page
permissions instead of trying to change their permissions to
PAGE_EXECUTE_READWRITE.

In order to write to memory-mapped region, we now try to use
copy on write support.  Note, that we call MapViewOfFileEx with
FILE_MAP_READ instead of FILE_MAP_COPY so that pages do not have
WRITECOPY permission right after MapViewOfFileEx call but we can
enable copy on write support later with VirtualProtect.

Unfortunately, Windows XP does not support PAGE_EXECUTE_WRITECOPY
page permission.  In this case we unmap files and map normal memory
with the same permissions instead.  This should normally happen only
when debugger sets breakpoint.  It doesn't matter whether the code
is mapped for most NaCl applications since we tightly restrict
the ways the code can be changed.

BUG= https://code.google.com/p/chromium/issues/detail?id=276912
TEST= run_gdb_mmap_test
R=bsy@google.com

Review URL: https://codereview.chromium.org/23579002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12129 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2