Unsandboxed IRT: Implement thread+futex interfaces, and some timer interfaces
This is enough to make run_thread_test and the IRT futex test pass.

Tweak the futex test so that it only tests a bad timeout address when
testing NaCl syscalls directly, not when testing the IRT futex
interface, otherwise the unsandboxed IRT crashes when trying to read
the timeout values for this test.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=3699
TEST=run_irt_futex_test_irt run_thread_test_irt

Review URL: https://codereview.chromium.org/27259002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12253 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2