Add a --retries=<N> flag, for flaky test environments.
The major change is that stdout and stderr need to be captured in memory, and only printed after sel_ldr completes successfully, to avoid output from a failed run to pollute the output.

Also, stdin has to be captured at the beginning and played back separately for each run.

On completion, the stdout from sel_ldr is printed, followed by the stderr.

BUG= none
R=dschuff@chromium.org, eliben@chromium.org

Review URL: https://codereview.chromium.org/42603002

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12323 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2