toolchain_build: Add source packages, incremental builds and LLVM build
This CL introduces the concept of source vs. build targets in the build
scripts. Source targets are not memoized and can be depended on by
build targets (which allows build targets to share source targets and to
depend on multiple sources). It also allows disabling syncing sources
via a flag. Targets can also now have an 'inputs' key which specifies file or
directory paths that get used as inputs in addition to the dependency outputs.

It also introduces the concept of incremental builds (currently enabled
via a flag). Incremental builds do not clobber the working directory before
they run, and Commands on build targets can be marked with the
skip_for_incremental attribute, which will cause them
to be skipped for incremental builds.

These concepts are used in a strightforward way to implement a build
of LLVM and the PNaCl driver scripts.

This is a re-land of r12429 (and r12430) with the additional change that
once.Run creates the output directory if it is absent.

R=bradnelson@google.com
BUG= https://code.google.com/p/nativeclient/issues/detail?id=3719

Review URL: https://codereview.chromium.org/82553011

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12441 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2